requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
ansi2html==1.8.0
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
bleach==6.0.0
bokeh==3.1.1
certifi==2023.5.7
charset-normalizer==3.1.0
chart-studio==1.1.0
click==8.1.3
colorcet==3.0.1
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dash==2.10.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datalookup==1.0.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastjsonschema==2.17.1
Flask==2.2.5
fonttools==4.39.4
frozenlist==1.4.0
geographiclib==2.0
geopy==2.3.0
greenlet==2.0.2
holoviews==1.16.0
hvplot==0.8.3
idna==3.4
ipykernel==6.23.1
ipython==8.13.2
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter-bokeh==3.0.7
jupyter-dash==0.4.2
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
linkify-it-py==2.0.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
multidict==6.0.4
nbformat==5.9.0
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.1
panel==1.0.3
param==1.13.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.14.1
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyct==0.5.0
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pyviz-comms==2.3.0
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
retrying==1.3.4
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
SQLAlchemy==2.0.15
SQLAlchemy-Utils==0.41.1
stack-data==0.6.2
tenacity==8.2.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.2
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.2
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
xyzservices==2023.5.0
yarl==1.9.2
